requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
et-xmlfile==1.1.0
idna==3.7
numpy==2.0.1
openpyxl==3.1.5
pandas==2.2.2
psutil==6.0.0
pyodbc==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
